Refactored Tests in their respective packages
